backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
apscheduler>=3.10.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
redis>=5.0.0
tenacity>=8.2.0
celery[redis]>=5.3.0
